Add Terrain editing through TerrainManipulator
Add TerrainTools editor window
Add wireframe and brush decal to MaterialRenderer
Change terrain and water renderers to use f32 heights
Move InputManager out of GameRenderer
Add MapUtil and use it in various places where we had duplicate code
Add PhysicsUtil which has a function for raycasting the world position of the cursor
Add DrawSphere3D function to DebugRenderer
Clean up old unused terrain related files
Updated Engine submodule

+++ Source/Game/Game/Rendering/Water/WaterRenderer.cpp
@@ -222,7 +222,7 @@ void WaterRenderer::Load(LoadDesc& desc)
             // The offsets here are flipped on purpose
             vertex.xCellOffset = yOffset;
             vertex.yCellOffset = xOffset;
-            vertex.height = f16(vertexHeight);
+            vertex.height = vertexHeight;
             vertex.uv = hvec2(uv);
 
             if (yOffset >= desc.endY || xOffset >= desc.endX)
